Added two new fields to the attires Model: ImageName of type String and AttireImage of type IFormFile.
Implemented the file/image upload functionality to the create attires post action method and update the corresponding view - create. Also update the delete attires post action method to remove the image from the images folder in wwwroot. Updated the details page to show the newly upload file. Note: Do not edit the file without updating the post edit action method in the attires controller. If you try to edit now, it will return a null value for file which will then result in null reference exception when the delete is invoked.

Please update the post action edit method in the attires controller to enable the edit functionality to record the correct File name.